Split out --list tests
